Bluetooth: OPP: Use long as file size instead of int
This patch fixes issue of problems in sending/receiving files
which cannot be accomodated in int type of Java. Instead of using
int, long data type is used, which can hold 64 bits data size in JAVA
which is suitable for storing file sizes in GB. Changes are done
in how updates are sent to progress bar which supports only int
by using the percentage of data transferred to update the progress.

CRs-Fixed: 583275

+++ src/com/android/bluetooth/opp/BluetoothOppTransferInfo.java
@@ -41,9 +41,9 @@ public class BluetoothOppTransferInfo {
 
     int mDirection;
 
-    int mTotalBytes;
+    long mTotalBytes;
 
-    int mCurrentBytes;
+    long mCurrentBytes;
 
     int mStatus;
